Add pubsub docs. Minor properties refactor
