+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/RecordParserProperties.java
@@ -45,9 +45,6 @@ public class RecordParserProperties implements ParserProperties {
 
     private boolean keepFiles = false;
 
-    @NotNull
-    private PersistProperties persist = new PersistProperties();
-
     @Override
     public Path getStreamPath() {
         return mirrorProperties.getDataPath().resolve(getStreamType().getPath());
@@ -56,26 +53,5 @@ public class RecordParserProperties implements ParserProperties {
     @Override
     public StreamType getStreamType() {
         return StreamType.RECORD;
-    }
-
-    @Data
-    public class PersistProperties {
-
-        private boolean claims = false;
-
-        private boolean contracts = true;
-
-        private boolean cryptoTransferAmounts = true;
-
-        private boolean files = true;
-
-        private boolean nonFeeTransfers = false;
-
-        private boolean systemFiles = true;
-
-        /**
-         * If configured the mirror node will store the raw transaction bytes on the t_transaction table
-         */
-        private boolean transactionBytes = false;
     }
 }
