+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/parser/record/entity/sql/SqlProperties.java
@@ -27,7 +27,7 @@ import org.springframework.validation.annotation.Validated;
 
 @Data
 @Validated
-@ConfigurationProperties("hedera.mirror.importer.parser.record.sql")
+@ConfigurationProperties("hedera.mirror.importer.parser.record.entity.sql")
 public class SqlProperties {
     /**
      * PreparedStatement.executeBatch() is called after every batchSize number of transactions from record stream file.
